Update setup.py method for retrieving the version of alluxio-py
The current method of specifying the version requires that the user
already have the `requests` module installed. This prevents adding
this module to any `requirements.txt` file.

Since other parts of the system actually use `alluxio.__version__`,
I have replaced how `setup.py` gets the current version. Using this
mechanism does this with minimal change to the rest of the tools.

I have also updated the project homepage to point to the correct
Github URL.